+++ backlog4s-core/src/main/scala/com/github/chaabaj/backlog4s/datas/User.scala
@@ -47,6 +47,6 @@ case class UpdateUserForm(
 )
 
 object UserT {
-  def id(value: Long): Id[User] = Id(value)
+  def id(value: Int): Id[User] = Id(value)
   val myself: Id[User] = Id(0)
 }
