+++ backlog4s-graphql/src/main/scala/com/github/chaabaj/backlog4s/graphql/schemas/AttachmentSchema.scala
@@ -13,7 +13,7 @@ object AttachmentSchema extends BacklogSchema[Unit, Attachment] {
         Field(
           "id",
           IntType,
-          resolve = _.value.id.value.toInt
+          resolve = _.value.id.value
         ),
         Field(
           "name",
